Move helper function to separate modules
