+++ lib/TileDBQuery/TileDBQuery.js
@@ -1,23 +1,4 @@
 "use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || function (mod) {
-    if (mod && mod.__esModule) return mod;
-    var result = {};
-    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
-    __setModuleDefault(result, mod);
-    return result;
-};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -43,18 +24,15 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getResults = exports.TileDBQuery = void 0;
+exports.TileDBQuery = void 0;
 const dataToQuery_1 = __importDefault(require("../utils/dataToQuery"));
-const bufferToData_1 = __importStar(require("../utils/bufferToData"));
 const capnpQueryDeSerializer_1 = __importDefault(require("../utils/capnpQueryDeSerializer"));
-const setNullables_1 = __importDefault(require("../utils/setNullables"));
 const v1_1 = require("../v1");
 const v2_1 = require("../v2");
-const getByteLengthOfDatatype_1 = __importDefault(require("../utils/getByteLengthOfDatatype"));
-const flatten_1 = __importDefault(require("../utils/flatten"));
 const getWriterBody_1 = __importDefault(require("../utils/getWriterBody"));
-const groupValuesByOffsets_1 = __importDefault(require("../utils/groupValuesByOffsets"));
-const isArrayOfArrays_1 = __importDefault(require("../utils/isArrayOfArrays"));
+const convertToArrayBufferIfNodeBuffer_1 = __importDefault(require("../utils/convertToArrayBufferIfNodeBuffer"));
+const getSizeInBytesOfAllAttributes_1 = __importDefault(require("../utils/getSizeInBytesOfAllAttributes"));
+const getResultsFromArrayBuffer_1 = __importDefault(require("../utils/getResultsFromArrayBuffer"));
 class TileDBQuery {
     constructor(params) {
         this.configurationParams = params;
@@ -82,7 +60,7 @@ class TileDBQuery {
                  * Axios in nodeJS environments casts the response to a Buffer object
                  * we convert it back to an ArrayBuffer if needed
                  */
-                const queryData = convertToArrayBufferIfNodeBuffer(queryResponse.data);
+                const queryData = convertToArrayBufferIfNodeBuffer_1.default(queryResponse.data);
                 const bufferWithoutFirstEightBytes = queryData.slice(8);
                 return capnpQueryDeSerializer_1.default(bufferWithoutFirstEightBytes);
             }
@@ -93,7 +71,7 @@ class TileDBQuery {
                  */
                 const errorIsABuffer = ((_c = (_b = e === null || e === void 0 ? void 0 : e.response) === null || _b === void 0 ? void 0 : _b.data) === null || _c === void 0 ? void 0 : _c.buffer) || ((_e = (_d = e === null || e === void 0 ? void 0 : e.response) === null || _d === void 0 ? void 0 : _d.data) === null || _e === void 0 ? void 0 : _e.length);
                 if (errorIsABuffer) {
-                    const errorArrayBuffer = convertToArrayBufferIfNodeBuffer(e.response.data);
+                    const errorArrayBuffer = convertToArrayBufferIfNodeBuffer_1.default(e.response.data);
                     const decodedMessage = new TextDecoder().decode(errorArrayBuffer);
                     throw new Error(decodedMessage);
                 }
@@ -117,7 +95,7 @@ class TileDBQuery {
              * Axios in nodeJS environments casts the response to a Buffer object
              * we convert it back to an ArrayBuffer if needed
              */
-            const queryData = convertToArrayBufferIfNodeBuffer(queryResponse.data);
+            const queryData = convertToArrayBufferIfNodeBuffer_1.default(queryResponse.data);
             /**
              * First 8 bytes of the response, contain a Uint64 number
              * which is the size of the response we skip it.
@@ -163,7 +141,7 @@ class TileDBQuery {
                  * Axios in nodeJS environments casts the response to a Buffer object
                  * we convert it back to an ArrayBuffer if needed
                  */
-                const queryData = convertToArrayBufferIfNodeBuffer(queryResponse.data);
+                const queryData = convertToArrayBufferIfNodeBuffer_1.default(queryResponse.data);
                 /**
                  * First 8 bytes of the response, contain a Uint64 number
                  * which is the size of the response we skip it.
@@ -208,7 +186,7 @@ class TileDBQuery {
         /**
          * Calculate the size of bytes of the attributes from the attributeBufferHeaders of the Query object.
          */
-        const numberOfBytesOfResults = getSizeInBytesOfAllAttributes(attributeHeaders);
+        const numberOfBytesOfResults = getSizeInBytesOfAllAttributes_1.default(attributeHeaders);
         /**
          * We get the last N bytes (N is the number of total bytes of the attributes), which contain
          * the results of all the attributes
@@ -219,7 +197,7 @@ class TileDBQuery {
             ...arraySchema.attributes,
         ];
         // Calculate results
-        const results = exports.getResults(resultsBuffer, attributeHeaders, mergeAttributesAndDimensions);
+        const results = getResultsFromArrayBuffer_1.default(resultsBuffer, attributeHeaders, mergeAttributesAndDimensions);
         return results;
     }
     throwError(e) {
@@ -230,7 +208,7 @@ class TileDBQuery {
          */
         const errorIsABuffer = ((_b = (_a = e === null || e === void 0 ? void 0 : e.response) === null || _a === void 0 ? void 0 : _a.data) === null || _b === void 0 ? void 0 : _b.buffer) || ((_d = (_c = e === null || e === void 0 ? void 0 : e.response) === null || _c === void 0 ? void 0 : _c.data) === null || _d === void 0 ? void 0 : _d.length);
         if (errorIsABuffer) {
-            const errorArrayBuffer = convertToArrayBufferIfNodeBuffer(e.response.data);
+            const errorArrayBuffer = convertToArrayBufferIfNodeBuffer_1.default(e.response.data);
             const decodedMessage = new TextDecoder().decode(errorArrayBuffer);
             throw new Error(decodedMessage);
         }
@@ -241,116 +219,4 @@ class TileDBQuery {
 }
 exports.TileDBQuery = TileDBQuery;
 exports.default = TileDBQuery;
-/**
- * If buffer is a NodeJS Buffer object we convert it back to an ArrayBuffer
- * @param buffer ArrayBuffer or Nodejs Buffer
- * @returns ArrayBuffer
- */
-function convertToArrayBufferIfNodeBuffer(buffer) {
-    if (buffer.buffer) {
-        return new Uint8Array(buffer).buffer;
-    }
-    return buffer;
-}
-/**
- * Add all buffers of an attribute
- * @param attr AttributeBufferHeader
- * @returns number of the total bytes of the attribute
- */
-const getAttributeSizeInBytes = (attr) => {
-    return (attr.fixedLenBufferSizeInBytes +
-        attr.varLenBufferSizeInBytes +
-        attr.validityLenBufferSizeInBytes);
-};
-/**
- * Calculate the total bytes of all the attributes
- * @param attributes
- * @returns number of the total bytes of all the attributes
- */
-const getSizeInBytesOfAllAttributes = (attributes) => attributes.reduce((accum, attr) => accum + getAttributeSizeInBytes(attr), 0);
-/**
- * Convert an ArrayBuffer to a map of attributes with their results
- * @param arrayBuffer The slice ArrayBuffer that contains the results
- * @param attributes
- * @param attributesSchema
- * @returns A map of attribute names with the results of every attribute
- */
-const getResults = (arrayBuffer, attributes, attributesSchema) => {
-    const data = {};
-    /**
-     * We start from the last attribute which is at the end of the buffer
-     */
-    attributes.reverse().reduce((offset, attribute) => {
-        const totalNumberOfBytesOfAttribute = getAttributeSizeInBytes(attribute);
-        if (!totalNumberOfBytesOfAttribute) {
-            data[attribute.name] = null;
-            return offset;
-        }
-        // If there are validityLenBufferSizeInBytes the attribute is nullable
-        const isNullable = !!attribute.validityLenBufferSizeInBytes;
-        // If there are varLenBufferSizeInBytes the attribute is varLengthSized
-        const isVarLengthSized = !!attribute.varLenBufferSizeInBytes;
-        const selectedAttributeSchema = getAttributeSchema(attribute.name, attributesSchema);
-        const negativeOffset = -1 * offset;
-        /**
-         * If attribute is varLengthSized, we ignore the first N bytes (where N = fixedLenBufferSizeInBytes)
-         * These first N bytes contain the offsets of the attribute, which is a uint64 array.
-         */
-        const start = negativeOffset -
-            totalNumberOfBytesOfAttribute +
-            (isVarLengthSized ? attribute.fixedLenBufferSizeInBytes : 0);
-        /**
-         * If attribute is isNullable we ignore the last N bytes (where N = validityLenBufferSizeInBytes)
-         * These last N bytes contain a uint8 array of zeros and ones, where every zero represents
-         * that in that index the attribute is null.
-         */
-        const ending = negativeOffset -
-            (isNullable ? attribute.validityLenBufferSizeInBytes : 0);
-        const end = ending ? ending : undefined;
-        let result = bufferToData_1.default(arrayBuffer.slice(start, end), selectedAttributeSchema.type);
-        let offsets = [];
-        if (isVarLengthSized) {
-            const BYTE_PER_ELEMENT = getByteLengthOfDatatype_1.default(selectedAttributeSchema.type);
-            const startOfBuffer = negativeOffset - totalNumberOfBytesOfAttribute;
-            const offsetsBuffer = arrayBuffer.slice(startOfBuffer, startOfBuffer + attribute.fixedLenBufferSizeInBytes);
-            /**
-             * Offsets are Uint64 numbers, buffer contains byte offsets though,
-             * e.g. if type of the attribute is an INT32 (4 bytes per number) and the offsets are [0, 3, 4]
-             * the buffer contains the offsets * bytes of the element instead of just the offsets [0, 3 * 4, 4 * 4] = [0, 12, 16]
-             */
-            const byteOffsets = Array.from(new BigUint64Array(offsetsBuffer));
-            // Convert byte offsets to offsets
-            offsets = byteOffsets.map((o) => Number(o) / BYTE_PER_ELEMENT);
-        }
-        if (isNullable) {
-            /**
-             * If attribute is Nullable, we get the last N bytes, cast it to uint8 array to get
-             * what is null.
-             */
-            const nullableArrayEnd = ending + attribute.validityLenBufferSizeInBytes;
-            const nullableArrayBuffer = arrayBuffer.slice(ending, nullableArrayEnd ? nullableArrayEnd : undefined);
-            const nullablesTypedArray = bufferToData_1.bufferToInt8(nullableArrayBuffer);
-            /**
-             * nullablesArray should be an array of zeros and ones (e.g. [0, 1, 1, 0])
-             * Every zero represents that in that specific index the attribute is NULL
-             */
-            const nullablesArray = Array.from(nullablesTypedArray);
-            result = setNullables_1.default(Array.from(result), nullablesArray, offsets);
-        }
-        // If result is a String slice the String by the offsets to make it an array
-        if (isVarLengthSized && typeof result === "string") {
-            result = groupValuesByOffsets_1.default([...result], offsets).map((s) => s.join(""));
-        }
-        data[attribute.name] = isArrayOfArrays_1.default(result) ? flatten_1.default(result) : result;
-        return offset + totalNumberOfBytesOfAttribute;
-    }, 0);
-    return data;
-};
-exports.getResults = getResults;
-/**
- * Get attribute data from attribute name, attribute data contains the type of the attribute (e.g. INT32, StringUTF8 etc)
- */
-const getAttributeSchema = (attrName, attributesSchema) => {
-    return attributesSchema.find((attr) => attr.name === attrName);
-};
 //# sourceMappingURL=TileDBQuery.js.map
