+++ src/v1/api.ts
@@ -165,6 +165,12 @@ export interface ArrayActivityLog {
      * @memberof ArrayActivityLog
      */
     query_ranges?: string;
+    /**
+     * stats for query
+     * @type {string}
+     * @memberof ArrayActivityLog
+     */
+    query_stats?: string;
 }
 /**
  * Object including array info and pagination metadata
